Make max file size a 64-bit int in order not to be limited to 4GB by the 32-bit int.
git-svn-id: svn+ssh://svn.cern.ch/reps/emu/trunk/TriDAS/emu@8348 4525493e-7705-40b1-a816-d608a930855b

+++ daq/writer/include/emu/daq/writer/RawDataFile.h
@@ -49,7 +49,7 @@ namespace emu { namespace daq { namespace writer {
   class RawDataFile{
 
   private:
-    uint32_t          maxFileSize_;	///< when exceeding this size [bytes], the file will be closed, and a new one opened
+    uint64_t          maxFileSize_;	///< when exceeding this size [bytes], the file will be closed, and a new one opened
     string            pathToFile_;	///< direcory where the file is to be written
     string            host_;          ///< host name
     string            appName_;	///< name of application producing the data
@@ -116,7 +116,7 @@ namespace emu { namespace daq { namespace writer {
     /// @param app name of application producing the data
     /// @param logger logger
     ///
-    RawDataFile(const uint32_t maxFileSize, 
+    RawDataFile(const uint64_t maxFileSize, 
 		const string pathToFile, 
 		const string host, 
 		const string appName,
